+++ examples/scan.rs
@@ -19,18 +19,18 @@ async fn delete_fake_data(client: &RedisClient) -> Result<(), RedisError> {
 
 #[tokio::main]
 async fn main() -> Result<(), RedisError> {
-  let config = RedisConfig::default();
-  let client = RedisClient::new(config, None, None);
-
+  let client = RedisClient::default();
   let _ = client.connect();
   let _ = client.wait_for_connect().await?;
   let _ = create_fake_data(&client).await?;
 
-  // build up a buffer of (key, value) pairs from pages (10 keys per page)
+  // build up a buffer of (key, value) pairs from pages (~10 keys per page)
   let mut buffer = Vec::with_capacity(COUNT as usize);
   let mut scan_stream = client.scan("foo*", Some(10), None);
 
-  while let Some(Ok(mut page)) = scan_stream.next().await {
+  while let Some(result) = scan_stream.next().await {
+    let mut page = result.expect("SCAN failed with error");
+
     if let Some(keys) = page.take_results() {
       // create a client from the scan result, reusing the existing connection(s)
       let client = page.create_client();
@@ -42,8 +42,8 @@ async fn main() -> Result<(), RedisError> {
       }
     }
 
-    // move on to the next page now that we're done reading the values
-    // or move this before we call `get` on each key to scan results in the background as fast as possible
+    // move on to the next page now that we're done reading the values. or move this before we call `get` on each key
+    // to scan results in the background as quickly as possible.
     let _ = page.next();
   }